ESI_AI_Helpdesk_Backend/app/utils/config.py
import os
from dotenv import load_dotenv
from sqlalchemy.ext.declarative import declarative_base
from sqlalchemy import create_engine
from sqlalchemy.orm import sessionmaker
from sqlalchemy.engine import URL

load_dotenv()

# openai configuration
OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
if not OPENAI_API_KEY:
    raise RuntimeError("OPENAI_API_KEY is not set")

DATABASE_URL = os.getenv("DATABASE_URL")
if not DATABASE_URL:
    raise RuntimeError("DATABASE_URL is not set")

# Database configuration
Base = declarative_base()

engine = create_engine(
    DATABASE_URL,
    pool_pre_ping=True,
    pool_size=10,
    max_overflow=20
)

# url = URL.create(
#     drivername=os.getenv("drivername"),
#     username=os.getenv("username"),
#     password=os.getenv("password"),
#     host=os.getenv("host"),
#     port=os.getenv("port"),
#     database=os.getenv("database"),
# )

# engine = create_engine(
#     url,
#     connect_args={"sslmode": "require"}  
# )

SessionLocal = sessionmaker(
    autocommit=False,
    autoflush=False,
    bind=engine
)

def get_db():
    db = SessionLocal()
    try:
        yield db
    finally:
        db.close()
